requirements.txt
attrs
certifi
charset-normalizer==2.0.7
execnet
idna==3.4
iniconfig
packaging
playwright==1.32.1
pluggy
py
pyee==9.0.4
pyparsing
pytest==7.2.0
allure-pytest
pytest-base-url
pytest-forked
pytest-playwright
pytest-xdist
python-slugify
requests==2.28.1
text-unidecode
toml
tomli
urllib3
websockets==10.4

jproperties
none~=0.1.1
configparser~=5.3.0
python-dotenv~=0.21.0
greenlet

boto3~=1.26.23
elasticsearch